+++ apps/vor/frontend/gatsby-browser.js
@@ -108,9 +108,11 @@ export const wrapPageElement = (params) => {
   window.localStorage.setItem(LANG_PREFERENCE_KEY, preferredLang)
   if (
     window.location.pathname + queryParams !== newUrl &&
-    window.location.pathname.replace(/\/$/, "") + queryParams !== newUrl
-  )
-    navigate(newUrl)
+    window.location.pathname.replace(/\/$/, "") + queryParams !== newUrl &&
+    !newUrl.endsWith("/404")
+  ) {
+    navigate(newUrl, {replace: true})
+  }
 
   return params.element
 }