+++ apps/vor/frontend/src/emotion.d.ts
@@ -1,7 +1,7 @@
 // You are also able to use a 3rd party theme this way:
 import "@emotion/react"
-import { Theme as ThemeUITheme } from "theme-ui"
+import { ObserfyTheme } from "./gatsby-plugin-theme-ui"
 
 declare module "@emotion/react" {
-  export interface Theme extends ThemeUITheme {}
+  export interface Theme extends ObserfyTheme {}
 }