+++ apps/vor/frontend/src/layouts/index.tsx
@@ -1,11 +1,19 @@
-import React, { FC } from "react"
 import { Global } from "@emotion/react"
+import React, { FC } from "react"
 import { QueryClient, QueryClientProvider } from "react-query"
-import { Box } from "theme-ui"
-import Layout from "../components/Layout/Layout"
+import { Box, useColorMode } from "theme-ui"
 import ErrorBoundary from "../components/ErrorBoundary/ErrorBoundary"
+import Layout from "../components/Layout/Layout"
 
-const queryClient = new QueryClient()
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      // refetchOnMount and refetchOnWindowFocus causes prefresh to fail and reload every once in a while.
+      // refetchOnMount: process.env.NODE_ENV !== "development",
+      // refetchOnWindowFocus: process.env.NODE_ENV !== "development",
+    },
+  },
+})
 
 // Used by gatsby-plugin-layout
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -22,16 +30,28 @@ const LayoutManager: FC<any> = ({ children, pageContext }) => (
   </QueryClientProvider>
 )
 
-const GlobalStyle: FC = () => (
-  <Global
-    styles={(theme) => ({
-      body: {
-        backgroundColor: theme.colors?.background ?? "",
-        minHeight: "100vh",
-        top: 0,
-      },
-    })}
-  />
-)
+const GlobalStyle: FC = () => {
+  const [mode] = useColorMode()
+  const customScroll = mode === "dark"
+
+  return (
+    <Global
+      styles={({ colors }) => ({
+        body: {
+          backgroundColor: colors.background,
+          minHeight: "100vh",
+          top: 0,
+          scrollbarColor: "dark",
+        },
+        "::-webkit-scrollbar": customScroll
+          ? {}
+          : { width: 8, backgroundColor: "#1c1c1c" },
+        "::-webkit-scrollbar-thumb": customScroll
+          ? {}
+          : { backgroundColor: "#3e3e3e", borderRadius: 9999 },
+      })}
+    />
+  )
+}
 
 export default LayoutManager